+++ ReleaseNotes.md
@@ -1,3 +1,8 @@
+# 3.0.388 (Caba series release 2)
+* Improved command line error reporting
+* `--inplace` option to instrument the assemblies in the input directory, having saved them to the output directory first; this avoids the manual copy-back step for `dotnet test` and the like
+* `--save` option in instrumenting mode plus `--collect` in `runner` mode; `--save` sets the instrumented assemblies to record coverage to disk, just as the `runner` mode does; then after doing whatever is required with the instrumented code, `runner --collect` will process the output just as if the operations had been run from within `AltCover runner`.  Note that if `--collect` is set, any arguments after a `-- ` are ignored.
+
 # 3.0.367-pre (Caba series release 1)
 * [BREAKING CHANGE] -- packaging the .net core as binaries -- so `dotnet AltCover.dll` rather than `dotnet run altcover.core.fsproj` to invoke.  This will be needed to do the MSBuild integration that's in the pipeline, where it isn't for a simple command-line tool
 * [BUGFIX] -- calculate branch exit visit count for `--opencover` (an oversight in 2.0.360)